Move setup_catalogs_order to setuphandlers
